Use list_sort from linux
Compared to the original merge sort that uses a top-down approach, the
bottom-up implementation in list_sort is more efficient in terms of the
stack usage.

Additionally, list_sort introduces a separate merge final function,
which handles the final merge and updates the prev pointer.

Moreover its functionality partially overlaps with the regular merge
function, it reduces the number of node access by maintaining the prev
pointer during the last merge step, thereby reducing the likelihood of
cache misses.

Change-Id: I32b464e05fae6eaa80f098a8b1171c9381dea60f